Add files via upload
5.15 (Hypotenuse Calculations) Define a function called hypotenuse that calculates the length
of the hypotenuse of a right triangle when the other two sides are given. Use this function in a program
to determine the length of the hypotenuse for each of the following triangles. The function
should take two arguments of type double and return the hypotenuse as a double. Test your program
with the side values specified in Fig. 5.18
  1-  3.0 4.0
  2-  5.0 12.0
  3-  8.0 15.0
    "Fig5.18"